Fix double slashes in certain forms submitted by InnerBrowser::submitForm()
The Codeception\Lib\InnerBrowser method getFormUrl() can under certain
conditions create a form URL where the path begins with double slashes.

Conditions for the error:
1. The page with the form lies on a first-level path, e.g.
example.com/register
2. The form has an action without a leading slash, e.g.
<form action="login">

This is valid and browsers understand what is meant.

On Unix the malformed action URL will look like this:
http://example.com//form-target

On Windows this is even worse because dirname() of a first-level directory
returns a backslash (\) instead of a forward slash.

See http://www.webmasterworld.com/php/3752500.htm

Thus a form URL with Codeception run on Windows will look like this:
http://example.com/\/form-target

Both malformed URLs can cause server errors, depending on the server
configuration mostly 404 or 500.

This commit adds a test that confirms the error and fixes it
in a localized way that shouldn't regress anything else.